# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Moustass API",
        "description": "API sécurisée pour l'application Moustass",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://www.moustass.com/api",
            "description": "Serveur local"
        }
    ],
    "paths": {
...
        "/auth/register": {
            "post": {
                "tags": [
                    "Authentification"
                ],
                "summary": "Inscription d'un nouvel utilisateur",
                "operationId": "ed42f392010c301892e640067141676e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "alice@moustass.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "MonMotDePasse1!"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Utilisateur créé avec succès"
                    },
                    "400": {
                        "description": "Données invalides"
                    }
                }
            }
        },
        "/auth/login": {
            "post": {
                "tags": [
                    "Authentification"
                ],
                "summary": "Connexion utilisateur",
                "operationId": "8cb7742ae98d0990acb2907c4e00bf24",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "alice@moustass.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "MonMotDePasse1!"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Connexion réussie"
                    },
                    "401": {
                        "description": "Identifiants invalides"
                    },
                    "403": {
                        "description": "Compte désactivé"
                    }
                }
            }
        },
        "/auth/me": {
            "get": {
                "tags": [
                    "Authentification"
                ],
                "summary": "Récupérer les infos de l'utilisateur connecté",
                "operationId": "a9c495788506b114956e3f509e5d90a2",
                "responses": {
                    "200": {
                        "description": "Informations utilisateur"
                    },
                    "401": {
                        "description": "Non authentifié"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/admin/backups/history": {
            "get": {
                "tags": [
                    "Backups"
                ],
                "summary": "Historique des sauvegardes",
                "operationId": "9f98aa034561bc528c6d713a6d5226db",
                "responses": {
                    "200": {
                        "description": "Liste des sauvegardes"
                    },
                    "401": {
                        "description": "Non authentifié"
                    },
                    "403": {
                        "description": "Accès refusé"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/admin/backups/full": {
            "post": {
                "tags": [
                    "Backups"
                ],
                "summary": "Créer une sauvegarde complète",
                "operationId": "b7929c855c22b0b2f57f07e340de61e2",
                "responses": {
                    "201": {
                        "description": "Backup complet créé avec succès"
                    },
                    "500": {
                        "description": "Erreur lors du backup"
                    },
                    "403": {
                        "description": "Accès refusé"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/admin/backups/incremental": {
            "post": {
                "tags": [
                    "Backups"
                ],
                "summary": "Créer une sauvegarde incrémentale",
                "operationId": "7c803bcce95416d4dc9fa1dda8a7acec",
                "responses": {
                    "201": {
                        "description": "Backup incrémental créé avec succès"
                    },
                    "400": {
                        "description": "Aucun backup complet trouvé"
                    },
                    "500": {
                        "description": "Erreur lors du backup"
                    },
                    "403": {
                        "description": "Accès refusé"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/admin/backups/restore": {
            "post": {
                "tags": [
                    "Backups"
                ],
                "summary": "Restaurer une sauvegarde",
                "operationId": "363f953697ac0df725417252e34112cd",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "backup_id"
                                ],
                                "properties": {
                                    "backup_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Base de données restaurée avec succès"
                    },
                    "404": {
                        "description": "Fichier de backup introuvable"
                    },
                    "500": {
                        "description": "Erreur lors de la restauration"
                    },
                    "403": {
                        "description": "Accès refusé"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}